+++ src/webhooks/types/events/directory_group_deleted.rs
@@ -14,6 +14,7 @@ mod test {
 
     use crate::directory_sync::{DirectoryGroupId, DirectoryId};
 
+    use crate::organizations::OrganizationId;
     use crate::webhooks::{Webhook, WebhookEvent, WebhookId};
     use crate::{RawAttributes, Timestamp, Timestamps};
 
@@ -30,6 +31,7 @@ mod test {
                   "object": "directory_group",
                   "id": "directory_group_01E1X5GPMMXF4T1DCERMVEEPVW",
                   "directory_id": "directory_01E1X194NTJ3PYMAY79DYV0F0P",
+                  "organization_id": "org_01EZTR6WYX1A0DSE2CYMGXQ24Y",
                   "idp_id": "02grqrue4294w24",
                   "name": "Developers",
                   "created_at": "2021-06-25T19:07:33.155Z",
@@ -58,6 +60,9 @@ mod test {
                         id: DirectoryGroupId::from("directory_group_01E1X5GPMMXF4T1DCERMVEEPVW"),
                         idp_id: "02grqrue4294w24".to_string(),
                         directory_id: DirectoryId::from("directory_01E1X194NTJ3PYMAY79DYV0F0P"),
+                        organization_id: Some(OrganizationId::from(
+                            "org_01EZTR6WYX1A0DSE2CYMGXQ24Y"
+                        )),
                         name: "Developers".to_string(),
                         timestamps: Timestamps {
                             created_at: Timestamp::try_from("2021-06-25T19:07:33.155Z").unwrap(),