+++ src/webhooks/types/events/directory_user_deleted.rs
@@ -16,6 +16,7 @@ mod test {
         DirectoryId, DirectoryUserEmail, DirectoryUserId, DirectoryUserState,
     };
 
+    use crate::organizations::OrganizationId;
     use crate::webhooks::{Webhook, WebhookEvent, WebhookId};
     use crate::{KnownOrUnknown, RawAttributes, Timestamp, Timestamps};
 
@@ -30,6 +31,7 @@ mod test {
               "data": {
                 "object": "directory_user",
                 "directory_id": "directory_01E1X194NTJ3PYMAY79DYV0F0P",
+                "organization_id": "org_01EZTR6WYX1A0DSE2CYMGXQ24Y",
                 "id": "directory_user_01E1X1B89NH8Z3SDFJR4H7RGX7",
                 "idp_id": "8931",
                 "first_name": "Veda",
@@ -78,6 +80,9 @@ mod test {
                         },
                         idp_id: "8931".to_string(),
                         directory_id: DirectoryId::from("directory_01E1X194NTJ3PYMAY79DYV0F0P"),
+                        organization_id: Some(OrganizationId::from(
+                            "org_01EZTR6WYX1A0DSE2CYMGXQ24Y"
+                        )),
                         username: Some("veda@foo-corp.com".to_string()),
                         emails: vec![DirectoryUserEmail {
                             primary: Some(true),